requirements.txt
# Core ML Libraries
tensorflow==2.15.0
scikit-learn==1.4.0
xgboost==2.0.3
numpy==1.24.3
pandas==2.1.4

# Model Optimization
tf2onnx==1.15.1
onnx==1.15.0
tensorflow-model-optimization==0.7.5
skl2onnx==1.14.1

# Monitoring and Metrics
prometheus-client==0.19.0
redis==5.0.1
mlflow==2.10.0
wandb==0.16.3

# Web Framework
Flask==3.0.2
gunicorn==21.2.0
flask-cors==4.0.0

# Testing
pytest==7.4.4
pytest-cov==4.1.0
coverage==7.4.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
joblib==1.3.2
scipy==1.12.0
typing-extensions==4.9.0

# Development Tools
black==24.1.1
pylint==3.0.3
pre-commit==3.6.0